Initial Draft
This is the initial draft of the school news app. Very disorganized, I know. Hopefully in later revisions it will become more clean.

Please also note that I have not gotten around to make a central `dbconnect.php` file, and so most if not all `.php` files require the info to connect to db. Later revisions will fix that.